changes names of Header to me able to use semantic Header, styles Header. Removes Home-link and makes El guacho clickable to indexpage. Removes celled grid in singe articles
