+++ src/components/SingleArticle.jsx
@@ -1,4 +1,4 @@
-import { Image, Grid, Header, Button, Segment } from "semantic-ui-react";
+import { Image, Grid, Header, Button, Segment, Container } from "semantic-ui-react";
 import React, { useState, useEffect } from "react";
 import { useParams, Link } from "react-router-dom";
 import Articles from "../modules/articles";
@@ -29,7 +29,8 @@ const SingleArticle = () => {
       {message ? (
         <p data-cy="error-message">{message}</p>
       ) : (
-        <Grid celled>
+        <Container>
+        <Grid >
           <Grid.Row>
             <Grid.Column width={13}>
               <Header data-cy="title">{article.title}</Header>
@@ -72,6 +73,7 @@ const SingleArticle = () => {
             </Segment>
           )}
         </Grid>
+        </Container>
       )}
     </>
   );
